move pkg/tool/analyzer to pkg/tool/analysis
